Update README, remove some TODO items.
Removed dupe filtering from the to-do list. A receive-only igate should not suppress duplicates, but rather leave it to the APRS-IS servers; they can utilize the duplicates too and this will give a more consistent implementation for the duplicate filtering. The duplicates can be extracted from the servers for network analysis, for example.
Removed "binary inspection", an iGate should pass the packet contents to the server as is, without making any special processing or de-encoding.